+++ tests/test.h
@@ -565,7 +565,6 @@ void test_consumer_close (rd_kafka_t *rk);
 void test_flush (rd_kafka_t *rk, int timeout_ms);
 
 void test_conf_set (rd_kafka_conf_t *conf, const char *name, const char *val);
-char *test_conf_get (const rd_kafka_conf_t *conf, const char *name);
 char *test_topic_conf_get (const rd_kafka_topic_conf_t *tconf,
                            const char *name);
 int test_conf_match (rd_kafka_conf_t *conf, const char *name, const char *val);